+++ tests/hypothesis_strategies/targets.py
@@ -9,6 +9,7 @@ from ..hypothesis_strategies.transformations import (
     chained_transformations,
     single_transformations,
 )
+from .metadata import measurable_metadata
 
 target_names = st.text(min_size=1)
 """A strategy that generates target names."""
@@ -20,7 +21,8 @@ def numerical_targets(draw: st.DrawFn):
     name = draw(target_names)
     transformation = draw(st.one_of(single_transformations, chained_transformations()))
     minimize = draw(st.booleans())
-    return NumericalTarget(name, transformation, minimize=minimize)
+    metadata = draw(measurable_metadata())
+    return NumericalTarget(name, transformation, minimize=minimize, metadata=metadata)
 
 
 choice_values = st.one_of(
@@ -34,7 +36,13 @@ def binary_targets(draw: st.DrawFn):
     """A strategy that generates binary targets."""
     name = draw(target_names)
     choices = draw(st.lists(choice_values, min_size=2, max_size=2, unique=True))
-    return BinaryTarget(name, success_value=choices[0], failure_value=choices[1])
+    target_metadata = draw(measurable_metadata())
+    return BinaryTarget(
+        name,
+        success_value=choices[0],
+        failure_value=choices[1],
+        metadata=target_metadata,
+    )
 
 
 targets = st.one_of([binary_targets(), numerical_targets()])