+++ tests/test_imports.py
@@ -50,7 +50,23 @@ def make_import_check(modules: Sequence[str], target: str) -> str:
     )
 
 
-@pytest.mark.parametrize("module", find_modules())
+_modules = find_modules()
+# TODO Remove fpsample workaround once fixed,
+#  see https://github.com/leonardodalinky/fpsample/issues/7
+_modules = [
+    param(
+        m,
+        marks=pytest.mark.xfail(
+            reason="fpsample is currently incompatible with Python 3.13", strict=True
+        ),
+    )
+    if sys.version_info[:2] == (3, 13) and "fpsample" in m
+    else m
+    for m in _modules
+]
+
+
+@pytest.mark.parametrize("module", _modules)
 def test_imports(module: str):
     """All modules can be imported without throwing errors."""
     importlib.import_module(module)
@@ -103,7 +119,11 @@ def test_lazy_loading(target: str, whitelist: Sequence[str]):
     """The target does not appear in the module list after loading BayBE modules."""
     all_modules = find_modules()
     assert (w in all_modules for w in whitelist)
-    modules = [i for i in all_modules if i not in whitelist]
+    modules = [m for m in all_modules if m not in whitelist]
+    if sys.version_info[:2] == (3, 13):
+        # Excluding fpsample modules here, as they are not available in Python 3.13
+        # TODO Remove once fixed, see https://github.com/leonardodalinky/fpsample/issues/7
+        modules = [m for m in modules if "fpsample" not in m]
     code = make_import_check(modules, target)
     python_interpreter = sys.executable
     result = subprocess.call([python_interpreter, "-c", code])